Split code up into multiple files and moved others to sub directory.
