+++ ConnectSdk/Search/IImagesSearch.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using GettyImages.Connect.Search.Entity;
 
 namespace GettyImages.Connect.Search
 {
