+++ ConnectSdk/Search/SearchImages.cs
@@ -1,10 +1,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using GettyImages.Connect.Search.Entity;
 
 namespace GettyImages.Connect.Search
 {
-    public class SearchImages : ApiRequest, IBlendedImagesSearch, ICreativeImagesSearch, IEditorialImagesSearch
+    public partial class SearchImages : ApiRequest, IBlendedImagesSearch, ICreativeImagesSearch, IEditorialImagesSearch
     {
         protected const string V3SearchImagesPath = "/search/images";
         private const string AgeOfPeopleKey = "age_of_people";
@@ -385,372 +386,6 @@ namespace GettyImages.Connect.Search
                 }
             }
             return this;
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithPage(int value)
-        {
-            return WithPage(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithPageSize(int value)
-        {
-            return WithPageSize(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithPhrase(string value)
-        {
-            return WithPhrase(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithSortOrder(string value)
-        {
-            return WithSortOrder(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithEmbedContentOnly(bool value)
-        {
-            return WithEmbedContentOnly(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithExcludeNudity(bool value)
-        {
-            return WithExcludeNudity(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithEventId(int value)
-        {
-            return WithEventId(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithResponseField(string value)
-        {
-            return WithResponseField(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithGraphicalStyle(GraphicalStyles value)
-        {
-            return WithGraphicalStyle(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithOrientation(Orientation value)
-        {
-            return WithOrientation(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithLicenseModel(LicenseModel value)
-        {
-            return WithLicenseModel(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithFileType(FileType value)
-        {
-            return WithFileType(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithKeywordId(int value)
-        {
-            return WithKeywordId(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithNumberOfPeople(NumberOfPeople value)
-        {
-            return WithNumberOfPeople(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithPrestigeContentOnly(bool value)
-        {
-            return WithPrestigeContentOnly(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithProductType(ProductType value)
-        {
-            return WithProductType(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithLocation(string value)
-        {
-            return WithLocation(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithAgeOfPeople(AgeOfPeople value)
-        {
-            return WithAgeOfPeople(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithComposition(Composition value)
-        {
-            return WithComposition(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithEthnicity(Ethnicity value)
-        {
-            return WithEthnicity(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithArtist(string value)
-        {
-            return WithArtist(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithCollectionCode(string value)
-        {
-            return WithCollectionCode(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithDateTo(string value)
-        {
-            return WithDateTo(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithDateFrom(string value)
-        {
-            return WithDateFrom(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithCollectionFilterType(CollectionFilter value)
-        {
-            return WithCollectionFilterType(value);
-        }
-
-        IBlendedImagesSearch IBlendedImagesSearch.WithSpecificPeople(string value)
-        {
-            return WithSpecificPeople(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithPage(int value)
-        {
-            return WithPage(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithPageSize(int value)
-        {
-            return WithPageSize(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithPhrase(string value)
-        {
-            return WithPhrase(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithSortOrder(string value)
-        {
-            return WithSortOrder(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithEmbedContentOnly(bool value)
-        {
-            return WithEmbedContentOnly(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithExcludeNudity(bool value)
-        {
-            return WithExcludeNudity(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithResponseField(string value)
-        {
-            return WithResponseField(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithGraphicalStyle(GraphicalStyles value)
-        {
-            return WithGraphicalStyle(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithOrientation(Orientation value)
-        {
-            return WithOrientation(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithLicenseModel(LicenseModel value)
-        {
-            return WithLicenseModel(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithFileType(FileType value)
-        {
-            return WithFileType(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithKeywordId(int value)
-        {
-            return WithKeywordId(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithNumberOfPeople(NumberOfPeople value)
-        {
-            return WithNumberOfPeople(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithPrestigeContentOnly(bool value)
-        {
-            return WithPrestigeContentOnly(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithProductType(ProductType value)
-        {
-            return WithProductType(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithLocation(string value)
-        {
-            return WithLocation(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithAgeOfPeople(AgeOfPeople value)
-        {
-            return WithAgeOfPeople(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithComposition(Composition value)
-        {
-            return WithComposition(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithArtist(string value)
-        {
-            return WithArtist(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithEthnicity(Ethnicity value)
-        {
-            return WithEthnicity(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithCollectionCode(string value)
-        {
-            return WithCollectionCode(value);
-        }
-
-        ICreativeImagesSearch ICreativeImagesSearch.WithCollectionFilterType(CollectionFilter value)
-        {
-            return WithCollectionFilterType(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithEditorialSegment(EditorialSegment segment)
-        {
-            EditorialSegments = EditorialSegments | segment;
-            return this;
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithPage(int value)
-        {
-            return WithPage(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithPageSize(int value)
-        {
-            return WithPageSize(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithPhrase(string value)
-        {
-            return WithPhrase(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithSortOrder(string value)
-        {
-            return WithSortOrder(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithEmbedContentOnly(bool value)
-        {
-            return WithEmbedContentOnly(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithExcludeNudity(bool value)
-        {
-            return WithExcludeNudity(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithResponseField(string value)
-        {
-            return WithResponseField(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithGraphicalStyle(GraphicalStyles value)
-        {
-            return WithGraphicalStyle(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithOrientation(Orientation value)
-        {
-            return WithOrientation(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithEventId(int value)
-        {
-            return WithEventId(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithFileType(FileType value)
-        {
-            return WithFileType(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithKeywordId(int value)
-        {
-            return WithKeywordId(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithNumberOfPeople(NumberOfPeople value)
-        {
-            return WithNumberOfPeople(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithProductType(ProductType value)
-        {
-            return WithProductType(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithLocation(string value)
-        {
-            return WithLocation(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithAgeOfPeople(AgeOfPeople value)
-        {
-            return WithAgeOfPeople(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithComposition(Composition value)
-        {
-            return WithComposition(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithArtist(string value)
-        {
-            return WithArtist(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithEthnicity(Ethnicity value)
-        {
-            return WithEthnicity(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithCollectionCode(string value)
-        {
-            return WithCollectionCode(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithDateTo(string value)
-        {
-            return WithDateTo(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithDateFrom(string value)
-        {
-            return WithDateFrom(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithCollectionFilterType(CollectionFilter value)
-        {
-            return WithCollectionFilterType(value);
-        }
-
-        IEditorialImagesSearch IEditorialImagesSearch.WithSpecificPeople(string value)
-        {
-            return WithSpecificPeople(value);
         }
 
         private void AddQueryParameter(string key, object value)
